src/requirements.txt
celery==5.3.6
Django==4.2.8
djangorestframework==3.14.0
redis==5.0.1